Bluetooth-OPP: Use long data type to contain receiving file size
If receiving file size is in GBs int type will not be able to contain
actual size and will lead to failure. So change to long instead of
int.

CRs-Fixed: 756089

+++ src/com/android/bluetooth/opp/BluetoothOppReceiveFileInfo.java
@@ -113,7 +113,7 @@ public class BluetoothOppReceiveFileInfo {
             try {
                 if (metadataCursor.moveToFirst()) {
                     hint = metadataCursor.getString(0);
-                    length = metadataCursor.getInt(1);
+                    length = metadataCursor.getLong(1);
                     mimeType = metadataCursor.getString(2);
                 }
             } finally {
